# docs/swagger.yaml
basePath: /v1
definitions:
  api.Agent:
    properties:
      createdAt:
        type: string
      id:
        type: string
      lastSeen:
        type: string
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      status:
        type: string
    type: object
...
  api.AgentStatus:
    properties:
      activeJobs:
        type: integer
      completedJobs:
        type: integer
      id:
        type: string
      isDraining:
        type: boolean
      lastSeen:
        type: string
      status:
        type: string
    type: object
  api.Artifact:
    properties:
      createdAt:
        type: string
      id:
        type: string
      jobId:
        type: string
      name:
        type: string
      path:
        type: string
      runId:
        type: string
      size:
        type: integer
      type:
        $ref: '#/definitions/api.ArtifactType'
    type: object
  api.ArtifactListResponse:
    properties:
      artifacts:
        items:
          $ref: '#/definitions/api.Artifact'
        type: array
      cursor:
        type: string
      hasMore:
        type: boolean
    type: object
  api.ArtifactType:
    enum:
    - pdf
    - pptx
    - zip
    - log
    type: string
    x-enum-varnames:
    - ArtifactTypePDF
    - ArtifactTypePPTX
    - ArtifactTypeZIP
    - ArtifactTypeLog
...
  api.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
    type: object
...
  api.RefreshRequest:
    properties:
      refreshToken:
        type: string
    type: object
...
host: localhost:3333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for the Agent Project Manager daemon. Orchestrates agent-based
    LLM workflows.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Agent Project Manager API
  version: "1.0"
paths:
...
  /agents/{agentId}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific agent
      parameters:
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Agent'
        "404":
          description: Agent not found
          schema:
            type: string
      summary: Get agent details
      tags:
      - agents
  /agents/{agentId}/drain:
    post:
      consumes:
      - application/json
      description: Stop an agent from accepting new work (drain mode)
      parameters:
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Agent not found
          schema:
            type: string
      summary: Drain agent
      tags:
      - agents
  /agents/{agentId}/resume:
    post:
      consumes:
      - application/json
      description: Resume an agent to accept new work
      parameters:
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: Agent not found
          schema:
            type: string
      summary: Resume agent
      tags:
      - agents
  /agents/{agentId}/status:
    get:
      consumes:
      - application/json
      description: Get status information for a specific agent
      parameters:
      - description: Agent ID
        in: path
        name: agentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AgentStatus'
        "404":
          description: Agent not found
          schema:
            type: string
      summary: Get agent status
      tags:
      - agents
  /artifacts:
    get:
      consumes:
      - application/json
      description: Get a paginated list of artifacts with optional filtering
      parameters:
      - description: Filter by job ID
        in: query
        name: jobId
        type: string
      - description: Filter by run ID
        in: query
        name: runId
        type: string
      - description: Maximum number of artifacts to return
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination
        in: query
        name: cursor
        type: string
      - description: Filter by artifact type (pdf|pptx|zip|log)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ArtifactListResponse'
      summary: List artifacts
      tags:
      - artifacts
  /artifacts/{artifactId}:
    delete:
      consumes:
      - application/json
      description: Delete an artifact
      parameters:
      - description: Artifact ID
        in: path
        name: artifactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "404":
          description: Artifact not found
          schema:
            type: string
      summary: Delete artifact
      tags:
      - artifacts
    get:
      consumes:
      - application/json
      description: Get metadata for a specific artifact
      parameters:
      - description: Artifact ID
        in: path
        name: artifactId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.Artifact'
        "404":
          description: Artifact not found
          schema:
            type: string
      summary: Get artifact metadata
      tags:
      - artifacts
  /artifacts/{artifactId}/download:
    get:
      consumes:
      - application/json
      description: Download the artifact file
      parameters:
      - description: Artifact ID
        in: path
        name: artifactId
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Artifact file
          schema:
            type: file
        "404":
          description: Artifact not found
          schema:
            type: string
      summary: Download artifact
      tags:
      - artifacts
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate and receive access token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Login
      tags:
      - auth
...
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh an access token using a refresh token
      parameters:
      - description: Refresh token request
        in: body
        name: refresh
        required: true
        schema:
          $ref: '#/definitions/api.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Refresh token
      tags:
      - auth
...
schemes:
- http
- https
swagger: "2.0"