# internal/api/handlers_artifacts.go
package api

import (
	"encoding/json"
	"net/http"

	"github.com/go-chi/chi/v5"
)

// handleListArtifacts handles GET /artifacts
// @Summary      List artifacts
// @Description  Get a paginated list of artifacts with optional filtering
// @Tags         artifacts
// @Accept       json
// @Produce      json
// @Param        jobId   query     string  false  "Filter by job ID"
// @Param        runId   query     string  false  "Filter by run ID"
// @Param        limit   query     int     false  "Maximum number of artifacts to return"
// @Param        cursor  query     string  false  "Cursor for pagination"
// @Param        type    query     string  false  "Filter by artifact type (pdf|pptx|zip|log)"
// @Success      200     {object}  ArtifactListResponse
// @Router       /artifacts [get]
func handleListArtifacts(w http.ResponseWriter, r *http.Request) {
	// TODO: Parse query parameters: jobId, runId, limit, cursor, type
	// jobId := r.URL.Query().Get("jobId")
	// runId := r.URL.Query().Get("runId")
	// limit := r.URL.Query().Get("limit")
	// cursor := r.URL.Query().Get("cursor")
	// artifactType := r.URL.Query().Get("type")

	// TODO: Implement artifact listing logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode([]interface{}{})
}

// handleGetArtifact handles GET /artifacts/{artifactId}
// @Summary      Get artifact metadata
// @Description  Get metadata for a specific artifact
// @Tags         artifacts
// @Accept       json
// @Produce      json
// @Param        artifactId  path      string  true  "Artifact ID"
// @Success      200         {object}  Artifact
// @Failure      404         {string}  string  "Artifact not found"
// @Router       /artifacts/{artifactId} [get]
func handleGetArtifact(w http.ResponseWriter, r *http.Request) {
	artifactId := chi.URLParam(r, "artifactId")

	// TODO: Implement artifact metadata retrieval
	_ = artifactId

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(map[string]interface{}{})
}

// handleDownloadArtifact handles GET /artifacts/{artifactId}/download
// @Summary      Download artifact
// @Description  Download the artifact file
// @Tags         artifacts
// @Accept       json
// @Produce      application/octet-stream
// @Param        artifactId  path      string  true  "Artifact ID"
// @Success      200         {file}    file    "Artifact file"
// @Failure      404         {string}  string  "Artifact not found"
// @Router       /artifacts/{artifactId}/download [get]
func handleDownloadArtifact(w http.ResponseWriter, r *http.Request) {
	artifactId := chi.URLParam(r, "artifactId")

	// TODO: Implement artifact file streaming
	_ = artifactId

	w.Header().Set("Content-Type", "application/octet-stream")
	w.WriteHeader(http.StatusOK)
}

// handleDeleteArtifact handles DELETE /artifacts/{artifactId}
// @Summary      Delete artifact
// @Description  Delete an artifact
// @Tags         artifacts
// @Accept       json
// @Produce      json
// @Param        artifactId  path      string  true  "Artifact ID"
// @Success      204         {string}  string  "No Content"
// @Failure      404         {string}  string  "Artifact not found"
// @Router       /artifacts/{artifactId} [delete]
func handleDeleteArtifact(w http.ResponseWriter, r *http.Request) {
	artifactId := chi.URLParam(r, "artifactId")

	// TODO: Implement artifact cleanup logic
	_ = artifactId

	w.WriteHeader(http.StatusNoContent)
}