# internal/api/handlers_auth.go
package api

import (
	"encoding/json"
	"net/http"
)

// handleLogin handles POST /auth/login
// @Summary      Login
// @Description  Authenticate and receive access token
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        credentials  body      LoginRequest  true  "Login credentials"
// @Success      200          {object}  LoginResponse
// @Failure      400          {string}  string  "Invalid request"
// @Failure      401          {string}  string  "Unauthorized"
// @Router       /auth/login [post]
func handleLogin(w http.ResponseWriter, r *http.Request) {
	var req LoginRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// TODO: Implement authentication logic
	// For now, return a placeholder response
	response := LoginResponse{
		AccessToken: "placeholder_token",
		ExpiresIn:   3600,
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// handleRefresh handles POST /auth/refresh
// @Summary      Refresh token
// @Description  Refresh an access token using a refresh token
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        refresh  body      RefreshRequest  true  "Refresh token request"
// @Success      200      {object}  LoginResponse
// @Failure      400      {string}  string  "Invalid request"
// @Failure      401      {string}  string  "Unauthorized"
// @Router       /auth/refresh [post]
func handleRefresh(w http.ResponseWriter, r *http.Request) {
	var req RefreshRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	// TODO: Implement token refresh logic
	response := LoginResponse{
		AccessToken: "refreshed_token",
		ExpiresIn:   3600,
	}

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)
	json.NewEncoder(w).Encode(response)
}

// handleLogout handles POST /auth/logout
// @Summary      Logout
// @Description  Logout and invalidate the current session
// @Tags         auth
// @Accept       json
// @Produce      json
// @Success      200  {string}  string  "OK"
// @Router       /auth/logout [post]
func handleLogout(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement logout logic (invalidate token, etc.)
	w.WriteHeader(http.StatusOK)
}